[#10] Add support for federation
Adds a script for standing up a number of federated zones together with
some iRODS CSCs. Running test_federation also works, but the tests fail
due to a bug in federation which dates back to 4.2.8.

In the effort to add this support, a number of new structures and
functions have been added:
  - Add context object which allows for convenient access to the Docker
client and Compose project information (including platform and database
flavor information by inspecting the Compose services)
  - Add "docker cp"-like utility for copying files and directories out
of a container onto the host machine
  - Added a number of convenience functions for installing and setting
up iRODS topologies via Compose projects so that common code can be
consolidated and reused more easily