[#30] copy out XML logs.
Introduces an option, --extra-logs-path, to the run_plugin_tests.py script.

The option specifies an auxiliary path to copy from the container out to the
host after tests are complete.  If not given, nothing is affected;  if given
without a value, the path to be copied out defaults to '/var/lib/irods/test-reports'
(the usual depot for .xml test output files generated via xmlrunner from the test
hook).