[#28] Added test_manager and test_runner
Adds two classes which manage context for running tests in containers.

The test_runner is an abstract base class that requires a single ethod
to be implemented by "concrete" test_runners which executes the test. In
this commit, the only test_runner implemented is the
`irods_python_suite`, which executes the tests from the iRODS python
test suite (i.e. /var/lib/irods/scripts/run_tests.py).

The test_manager manages a number of test_runners so that tests can be
distributed to the test_runners and then executed in parallel on the
available pool of zones.