[#49] Add new platform support (main only)
The following platforms were added, but are not completely working yet:
- almalinux:8
- debian:11
- ubuntu:20.04

debian:11 is not working all the way because you can't run services
inside of the container (i.e. rsyslog). The same problem plagues
centos:7 and almalinux:8.

almalinux:8 and debian:11 do not have .tar.gz files available for any
MySQL ODBC drivers, they are only available as packages. This should be
addressed in the future.