[#68] Rework stream_output arg in execute_command
stream_output used to be a boolean value which would force the global
log level to INFO and stream the output of the command being executed.
We now accept (and default to) None to indicate that the configured log
level should dictate whether or not the output should be streamed. If
the output is explicitly set to stream, it will always stream at log
level ERROR (to ensure that it always shows up regardless of configured
log level). If the output is explicitly set not to stream, it will never
stream the output, even if the configured log level is at its highest
level.