[#71] Fix options for test_runners
The run() method for the test_manager now has a **kwargs parameter which
is ultimately forwarded to the test_runner's execute_test() method. The
python test suite runner has an options argument which is used directly
in the run_tests.py parameters. The unit test runner has an option to
configure the reporter type (--reporter for Catch2).