[#94] Modify core.re for SSL configuration
In 4.3.0, the default value returned by acPreConnect is now
CS_NEG_REFUSE, which will cause an SSL-enabled client to fail in the
SSL negotiation stage. This change adds the modification of acPreConnect
in core.re as part of the SSL configuration step enacted by --use-ssl.

Also stops the server before configuring SSL and simply starts it after
it's done rather than restarting at the end. This has proven to cause
problems in the server recently.