[#49] Start rsyslogd after setting up iRODS
Adds modified rsyslog.conf files to certain database flavors in order to
ensure that messages from the iRODS server are being processed and
logged to the log file. The conf file is copied into the Docker IMAGES
and will not be overwritten when rsyslog is installed/started.

Also adds a step to setting up the iRODS server to ensure that rsyslog
is indeed running before finally bringing up the server for use.