[#107] enable powertools for running plugin tests on almalinux
The irods_netcdf plugins are one instance where an extra set of
packages called PowerTools need to be enabled for install as
 - dependency targets for to the plugin's shared libraries.
 - runtime functionality for running the plugin tests.

EPEL used to serve this purpose on RHEL-7-ish, but is no longer
sufficient by itself.

Note as well that on RHEL-8 compatible platforms, "yum" and "dnf"
commands are symbolic links to a common script, "/usr/bin/dnf-3".