[#110] Add release Dockerfiles
Adds a file called release.Dockerfile for every platform. These
Dockerfiles install the iRODS server, database plugin, and iCommands
with a version specified by build-arg ${irods_package_version}. This
defaults to 4.3.0 (and whatever suffix is required for each platform).
This value will be updated to whatever the latest release is whenever a
release occurs.

Each Compose project now considers two environment variables which are
used as build-args:

 - dockerfile: Selects which Dockerfile to build. If this environment
variable is not set, Compose sets it to an empty string, and this
results in the default value of "Dockerfile" for the dockerfile
parameter. The only other valid option right now is release.Dockerfile.

 - irods_package_version: Specifies which version of iRODS to install
when building the image. This build-arg is unused for the default
Dockerfiles because they do not install anything until runtime. For
release.Dockerfiles, this string will be used directly in the
installation command when building the Docker image.

The environment variables are set automatically based on the options
used with the various testing environment scripts. A new option has been
added to the package version CLI options:

 - --use-static-image: This option specifies that the static Docker
image should be used instead of the image which requires that packages
be installed at runtime. This must be used in conjunction with
--irods-package-version which is incompatible with
--irods-package-directory.