[#127] Replace xmlrunner with unittest-xml-reporting
The pip package for xmlrunner should be installed via the
unittest-xml-reporting package instead of xmlrunner. xmlrunner is
supported for python2, so the package will remain the same for platforms
which support iRODS versions which use python2.

Also, aligned release Dockerfiles with regular Dockerfiles such that
build caches can be leveraged.