[#140] Implement better negotiation keys
Implements better negotiation key generation both in-zone and across
federation. make_negotiation_key's signature has been modified to take a
simple string as a "prefix" to identify the negotiation key and
underscores will pad out the string to be 32 characters long.

Federation negotiation keys are now generated by putting the local zone
name and the remote zone name in a list, alphabetizing them, and joining
them with an underscore. For instance, if your zone names are zoneA and
zoneB, the generated negotiation key would be:

  zoneA_zoneB_____________________

Also makes generating JSON files in the containers a bit cleaner by
supporting indentation and sorting the keys for readability.