[irods/irods#6482] Creates zonehopper#<remote_zone> for federation testing.
The federation test suite in irods/irods has limited capabilities.
Specifically, it cannot configure the remote zone for various
situations.

This commit causes the testing environment to create a new rodsuser in
the remote zone. The user is given a password which allows the
federation tests to connect to the remote zone as the new user. By doing
this, tests will be able to trigger delay rules across zones as well as
other things.

This commit guarantees that two different users share the same username,
but different zone names. Some tests require this to verify the system
operates as expected.