[154] Adjusted charset and collation settings for 4.2 and MySQL 8.
The iRODS 4.2 series uses a database schema which isn't compatible with
MySQL 8. This commit fixes that by disabling support for UTF-8 in the
MySQL 8 database.

This change is fine for test purposes as long as the data being placed
in the database satisfies the requirements defined by the latin1 charset
and latin1_general_cs collation scheme.