[#52] Add parallelization for topology tests
This change allows topology tests to run in parallel.

In order for this to work, the hostnames for the catalog consumers in
each zone on which the executors are running tests must be passed as
part of the base option string used for running the tests. The
test_runners are not aware of the zone in which they are executing
tests, only the container on which the tests are being executed.

Therefore, the run_topology_tests.py script is responsible for working
out the options strings for all of the executors and passing the options
strings to the interface for running tests. The options string is now a
list of options strings which will be given to each test_runner in
order when the tests are run.