Use only the last part of the image name
The scripts for building the images in some places assume the image name
retrieved via the "compose" library is e.g.

    ubuntu-2204-postgres-148_irods-catalog-provider

and not

    docker.io/library/ubuntu-2204-postgres-148_irods-catalog-provider

Trim the name in these cases by splitting on '/' and using the last part.

Noticed while using the testing environment with "podman" instead of
"docker" but perhaps more generally encountered with having multiple
registries configured.