When first setting up the server the service account may not yet exist
In this case trying to stop the server running as this account will
obviously fail until the setup has created this very account. Allow for
this case.

The specific error message to check for should perhaps not include
hardcoded "irods" as username, but is used in the present release in a
fixed way anyway and with few call-sites, and it makes the code somewhat
more self-explaining.